--- src/Shop.java
@@ -1,58 +1,0 @@
-import java.util.List;
-
-public class Shop {
-    private String name;
-    private String address;
-    private List<ProductDetails> products;
-    private String[] type;
-    private List<Service> services;
-    private List<Employee> employees;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getAddress() {
-        return address;
-    }
-
-    public void setAddress(String address) {
-        this.address = address;
-    }
-
-    public List<ProductDetails> getProducts() {
-        return products;
-    }
-
-    public void setProducts(List<ProductDetails> products) {
-        this.products = products;
-    }
-
-    public String[] getType() {
-        return type;
-    }
-
-    public void setType(String[] type) {
-        this.type = type;
-    }
-
-    public List<Service> getServices() {
-        return services;
-    }
-
-    public void setServices(List<Service> services) {
-        this.services = services;
-    }
-
-    public List<Employee> getEmployees() {
-        return employees;
-    }
-
-    public void setEmployees(List<Employee> employees) {
-        this.employees = employees;
-    }
-}
