+++ src/shopilingus/Location.java
@@ -1,0 +1,4 @@
+package shopilingus;
+
+public class Location {
+}
