+++ src/shopilingus/ShoppingCenter.java
@@ -1,0 +1,3 @@
+public class ShoppingCenter {
+
+}
